requirements.txt
xmltodict3==0.0.4
scikit-learn==0.24.1
pandas==1.2.3
torch==1.8.1
transformers==4.4.2
tqdm==4.59.0
pyyaml==5.4.1
fire==0.4.0
wandb==0.10.25
